+++ lib/mirrmaid/manager.py
@@ -19,6 +19,7 @@ import os
 import pwd
 import sys
 from argparse import ArgumentParser
+from signal import SIGHUP, SIGINT, SIGQUIT, SIGTERM, signal
 from traceback import format_exc
 
 import yaml
@@ -26,7 +27,7 @@ from doubledog.config.sectioned import DefaultConfig, InvalidConfiguration
 
 from mirrmaid.config import MirrmaidConfig, MirrorConfig, MirrorsConfig
 from mirrmaid.constants import *
-from mirrmaid.exceptions import SynchronizerException
+from mirrmaid.exceptions import SignalException, SynchronizerException
 from mirrmaid.logging.handlers import ConsoleHandler
 from mirrmaid.logging.kludge import race_friendly_rotator
 from mirrmaid.logging.summarizer import LogSummarizingHandler
@@ -43,6 +44,7 @@ class MirrorManager(object):
         self.args = args
         self._drop_privileges()
         self.parser = None
+        self._workers = None
         self._init_logger()
 
     def _config_logger(self):
@@ -69,6 +71,12 @@ class MirrorManager(object):
         else:
             os.environ[RSYNC_PROXY] = proxy
             _log.debug('will proxy rsync through %r', proxy)
+
+    def _config_signal_handler(self):
+        """Register a signal handler for graceful shutdowns."""
+        for signal_ in [SIGHUP, SIGINT, SIGQUIT, SIGTERM]:
+            _log.debug('setting trap for signal %r', signal_)
+            signal(signal_, self._signal_handler)
 
     def _config_summarizer(self):
         handler = LogSummarizingHandler(self.mirrmaid_conf)
@@ -164,13 +172,25 @@ class MirrorManager(object):
         self.default_conf = DefaultConfig(self.args.config_filename)
         self.mirrors_conf = MirrorsConfig(self.args.config_filename)
         _log.debug('enabled mirrors: %r', self.mirrors_conf.mirrors)
+        self._config_signal_handler()
+        self._workers = []
         for mirror in self.mirrors_conf.mirrors:
             _log.debug('processing mirror: %r', mirror)
             worker = Synchronizer(
                 self.default_conf,
                 MirrorConfig(self.args.config_filename, mirror)
             )
-            worker.run()
+            self._workers.append(worker)
+            worker.start()
+
+    def _signal_handler(self, signal_, _):
+        """React to signals to bring about graceful shutdown of workers."""
+        worker: Synchronizer
+        _log.debug('caught signal %r; halting all workers', signal_)
+        for worker in self._workers:
+            worker.stop()
+        _log.debug('all workers stopped or killed; shutting down')
+        raise SignalException(f'caught signal {signal_!r}')
 
     def run(self):
         # noinspection PyBroadException
@@ -182,8 +202,8 @@ class MirrorManager(object):
         except SynchronizerException as e:
             _log.critical(e)
             self._exit(os.EX_OSERR, e)
-        except KeyboardInterrupt:
-            _log.error('interrupted via SIGINT')
+        except (KeyboardInterrupt, SignalException) as e:
+            _log.error('interrupted via %s', e)
             self._exit(os.EX_OSERR)
         except SystemExit:
             pass  # presumably already handled